+++ package.json
@@ -11,9 +11,9 @@
   "scripts": {
     "lint": "eslint --quiet '**/*.{ts,js}'",
     "lint:fix": "npm run lint -- --fix",
-    "test": "jest --selectProjects Dev",
-    "test:ci": "jest --selectProjects CI",
-    "test:clean": "rimraf ./tests/data/*",
+    "test": "lerna run test",
+    "test:ci": "lerna run test:ci",
+    "test:clean": "lerna run test:clean",
     "clean": "lerna run clean",
     "build": "lerna run build",
     "package": "lerna run package",
@@ -23,30 +23,27 @@
     "playground": "npm start --workspace=@cere-ddc-sdk/playground"
   },
   "dependencies": {
-    "@types/jest": "^27.4.0",
     "@types/node": "^18.17.1",
     "clean-publish": "^4.2.0",
     "eslint": "^8.54.0",
-    "internal-ip": "6.2.0",
-    "jest": "^29.7.0",
-    "lerna": "^7.2.0",
+    "lerna": "^7.4.2",
     "microbundle": "^0.15.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.1.0",
     "rimraf": "^5.0.5",
-    "testcontainers": "^8.10.1",
-    "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
     "typescript": "^4.5.5",
-    "wait-port": "^1.1.0",
-    "ws": "^8.14.2"
+    "wait-port": "^1.1.0"
   },
-  "workspaces": [
-    "packages/*",
-    "playground"
-  ],
-  "overrides": {
-    "@polkadot/types": "9.4.1"
+  "workspaces": {
+    "packages": [
+      "packages/*",
+      "playground",
+      "tests"
+    ],
+    "nohoist": [
+      "playground/**"
+    ]
   }
 }
