+++ tests/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@cere-ddc-sdk/tests",
+  "version": "2.0.0-rc.10",
+  "description": "Cere DDC SDK tests",
+  "private": true,
+  "module": "./index.ts",
+  "scripts": {
+    "test": "jest --selectProjects Dev",
+    "test:ci": "jest --selectProjects CI",
+    "test:clean": "rimraf ./data/*"
+  },
+  "dependencies": {
+    "@types/jest": "^27.4.0",
+    "internal-ip": "6.2.0",
+    "jest": "^29.7.0",
+    "testcontainers": "^8.10.1",
+    "ts-jest": "^29.1.1",
+    "ws": "^8.14.2"
+  }
+}
